+++ cppForSwig/BIP150_151.h
@@ -119,7 +119,7 @@ private:
    std::array<uint8_t, BIP151PRVKEYSIZE> sessionID_{}; // Session ID
    std::array<uint8_t, BIP151PRVKEYSIZE*2> hkdfKeySet_{}; // K1=Payload, K2=Data size
    btc_key genSymECDHPrivKey_; // Prv key for ECDH deriv. Delete ASAP once used.
-   uint32_t bytesOnCurKeys_ = 0; // Bytes ctr for when to switch
+   size_t bytesOnCurKeys_ = 0; // Bytes ctr for when to switch
    BIP151SymCiphers cipherType_ = BIP151SymCiphers::INVALID;
    uint32_t seqNum_ = 0;
    bool encinit_ = false;
@@ -161,7 +161,7 @@ public:
    bool handshakeComplete() const {
       return (encinit_ == true && encack_ == true);
    }
-   uint32_t getBytesOnCurKeys() const { return bytesOnCurKeys_; }
+   size_t getBytesOnCurKeys() const { return bytesOnCurKeys_; }
    void setOutgoing() { isOutgoing_ = true; }
    bool getOutgoing() const { return isOutgoing_; }
    bool getSeqNum() const { return seqNum_; }
